Module solana_bpf_loader_program::syscalls

Modules

cpi 🔒

logging 🔒

mem_ops 🔒

sysvar 🔒

Macros

register_feature_gated_function 🔒

Structs

SyscallAbort

Abort syscall functions, called when the SBF program calls abort() LLVM will insert calls to abort() if it detects an untenable situation, abort() is not intended to be called explicitly by the program. Causes the SBF program to be halted immediately

SyscallAllocFree

Dynamic memory allocation syscall called when the SBF program calls sol_alloc_free_(). The allocator is expected to allocate/free from/to a given chunk of memory and enforce size restrictions. The memory chunk is given to the allocator during allocator creation and information about that memory (start address and size) is passed to the VM to use for enforcement.

SyscallAltBn128

alt_bn128 group operations

SyscallBigModExp

Big integer modular exponentiation

SyscallBlake3

SyscallCreateProgramAddress

Create a program address

SyscallCurveGroupOps

SyscallCurveMultiscalarMultiplication

SyscallCurvePointValidation

SyscallGetClockSysvar

Get a Clock sysvar

SyscallGetEpochScheduleSysvar

Get a EpochSchedule sysvar

SyscallGetFeesSysvar

Get a Fees sysvar

SyscallGetProcessedSiblingInstruction

Get a processed sigling instruction

SyscallGetRentSysvar

Get a Rent sysvar

SyscallGetReturnData

Get return data

SyscallGetStackHeight

Get current call stack height

SyscallInvokeSignedC

Cross-program invocation called from C

SyscallInvokeSignedRust

Cross-program invocation called from Rust

SyscallKeccak256

SyscallLog

Log a user’s info message

SyscallLogBpfComputeUnits

Log current compute consumption

SyscallLogData

Log data handling

SyscallLogPubkey

Log 5 64-bit values

SyscallLogU64

Log 5 64-bit values

SyscallMemcmp

memcmp

SyscallMemcpy

memcpy

SyscallMemmove

memmove

SyscallMemset

memset

SyscallPanic

Panic syscall function, called when the SBF program calls ’sol_panic_()` Causes the SBF program to be halted immediately

SyscallSecp256k1Recover

secp256k1_recover

SyscallSetReturnData

Set return data

SyscallSha256

SHA256

SyscallTryFindProgramAddress

Create a program address

Enums

SyscallError

Error definitions

Constants

MAX_SIGNERS

Maximum signers

Functions

consume_compute_meter 🔒

create_loader

translate 🔒

translate_and_check_program_address_inputs 🔒

translate_slice 🔒

translate_slice_inner 🔒

translate_slice_mut 🔒

translate_string_and_do 🔒

Take a virtual pointer to a string (points to SBF VM memory space), translate it pass it to a user-defined work function

translate_type 🔒

translate_type_inner 🔒

translate_type_mut 🔒

Type Definitions

Error 🔒